+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 * support Java 9 and 10 
   * renamed World.locateClasspathItem() to locateClasspathEntry() to make Java 9 upgrade problems explicit: 
     localteClasspathEntry does not locate modules
+  * fixed Serializer.serializeChildren to *never* format the output
 
+* added Serializer.serialize(node, format) 
 
 ### 3.1.7 (2018-04-25)